refactor: Phase 5 — プレゼンテーション層のリファクタリング
- hooks.ts の各フックを hooks/ ディレクトリの個別ファイルに分割
  - useInterval, useKeyboard, useIdleTimer を個別ファイルに移動
  - hooks.ts は後方互換性のための再エクスポートファイルに変換
- use-game-controls.ts の setTimeout を useSafeTimeout に置換（メモリリーク防止）
- use-skill-system.ts の setTimeout を useSafeTimeout に置換（メモリリーク防止）
